#8 - #2에서 구현한 코드에서 Calculator 클래스 분리
- main 메소드 안에서 처리하던 연산 처리를 Calculator 클래스에서 수행하도록 코드를 분리했다.
- 기존 코드와 동일한 동작하는지 확인 완료
